Adds a character repository and controller and moves the getCharacters and assignCharacter functions over. Renames the endpoints to match.
